Refactored ResultSet to be a member of QueryBuilder, rather than an inheritor
