+++ src/pds/peppi/products.py
@@ -1,9 +1,9 @@
 """Main class of the library in this module."""
 from .client import PDSRegistryClient
-from .result_set import ResultSet
+from .query_builder import QueryBuilder
 
 
-class Products(ResultSet):
+class Products(QueryBuilder):
     """Use to access any class of planetary products via the PDS Registry API.
 
     This class is both a :class:`.query_builder.QueryBuilder` which carries methods to subset the products
@@ -18,4 +18,4 @@ class Products(ResultSet):
         client : PDSRegistryClient
             Client defining the connexion with the PDS Search API
         """
-        ResultSet.__init__(self, client)
+        super().__init__(client)
